requirements.txt
tensorflow==2.18.0        # Core ML framework
keras==3.8.0              # High-level API for TensorFlow
numpy                     # For numerical operations
pillow                    # For image processing
matplotlib                # For visualizations 
streamlit==1.28.1         # Web app framework
scikit-learn              # For metrics like accuracy/confusion matrix
opencv-python==4.8.1.78   # for image manipulation
